Simplify license handling in Authorization class
Removed the conditional block for unrecognized licenses, eliminating the special case where the license is verified by the store but not known to us. This used to be possible as a result of an app communications failure at an inopportune moment but now the web service handles license acknowledgment it is no longer necessary.